websockets_server/requirements.txt
kafka-python==2.0.1
websockets==8.1